chore: 로그인 한 사용자 조회 api 경로 변경
